Block popups through `onCreatedNavigationTarget`.
Finally! This was discovered possible thanks to an answer on my SO
question (see https://stackoverflow.com/questions/46407042). Now,
for the first time in the life of this extension, I am able to
correctly and reliably block popup tabs and windows created by any
blacklisted site! Huge improvement over the reliability of the
extension.