BG: prevent Firefox from destroying options on save.
Recently Firefox decided to start treating objects with custom getters
and setters differently from others. This destroys the `options`
object every single time it is saved in the storage, because the
object has custom getters and setters defined by VueJS. Those
getters/setters should be transparent, and in fact cause no problem in
Google Chrome, but apparently Firefox is not using the getters when
saving the options in the storage, and this results in an empty object
being stored. The JSON.parse(JSON.stringify()) hack is just a silly
way to quickly "clone" the object getting rid of all the custom
getters and setters.